Add `conf.tea`, fix API functions, other fixes
Simplified a bit `init.tea`, thanks to the introduction of spreading.

Added `micro.gfx.setMaxFps`, as well as actually implementing some other API functions.

The idea would be to not use GLFW at all and write the handling ourselves. It will need a bit of work, but then we won't depend from any external library.

Well, there's also GLEW, to access OpenGL extension function, which also needs to be written in.